% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running.R
\name{cdsknn}
\alias{cdsknn}
\title{CDSKNN clustering}
\usage{
cdsknn(
  dat = NULL,
  partition_count = 300,
  batch_size = 500,
  outlier_det = TRUE,
  outlier_methods = "md",
  outlier_q = 0.2,
  min_n = 200,
  num_init = 10,
  max_iters = 1000,
  cluster_method = "louvain",
  resolution = 1,
  knn_range = c(3:10),
  iter = 20,
  is_weight = TRUE,
  assess_index = "Calinski_Harabasz",
  res_range = seq(0.2, 3, 0.2),
  new_cluster_method = "louvain",
  python_path = "/usr/bin/python3",
  seed = 723
)
}
\arguments{
\item{dat}{expression matrix, row is cell, col is feature}

\item{partition_count}{the number of partitions}

\item{batch_size}{the number of threads}

\item{outlier_det}{whether running outlier detection}

\item{outlier_methods}{outlier detection methods, md or ed.
"md" utilizes the Mahalanobis distance for detection,
while "ed" treats points that are far from the center proportionally as outliers.}

\item{outlier_q}{the proportion of outliers you want to filter.
When outlier_methods is ed, this parameter takes effect.}

\item{min_n}{the minimum number of data points for outlier detection.}

\item{num_init}{the proportion of outliers that need to be removed}

\item{max_iters}{the maximum iterations for kmeans}

\item{cluster_method}{louvain or leiden in resampling}

\item{resolution}{clustering parameters settings in resampling}

\item{knn_range}{the range of the number of neighbors in the KNN graph structure}

\item{iter}{the number of iterations in resampling}

\item{is_weight}{Whether to use distance weights when constructing the KNN graph}

\item{assess_index}{evaluation index used to select the optimal KNN graph structure}

\item{res_range}{resolution range in clustering}

\item{new_cluster_method}{louvain or leiden}

\item{python_path}{python path to be used}

\item{seed}{random seed}
}
\description{
This function is a wrapper that executes all steps of
CDSKNN clustering analysis in one go.
}
