% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{CreateKNN}
\alias{CreateKNN}
\title{Create knn-graph}
\usage{
CreateKNN(dat, k, is_weight = FALSE, python_path = "/usr/bin/python3")
}
\arguments{
\item{dat}{row is cell, col is feature}

\item{k}{the number of nearest neighbors}

\item{is_weight}{whether to use distance weights when constructing the KNN graph}

\item{python_path}{python path to be used}
}
\description{
This function is used to running Louvain clustering
}
