% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{faiss_ann}
\alias{faiss_ann}
\title{faiss_ann}
\usage{
faiss_ann(dat, k, python_path)
}
\arguments{
\item{dat}{used to create knn graph}

\item{k}{the result from mbkmeans function}

\item{python_path}{python path to be used}
}
\description{
The invocation interface of faiss in Python.
}
