% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResult.R
\name{GetMapClusterLabel}
\alias{GetMapClusterLabel}
\title{Map the K-means results to all cells}
\usage{
GetMapClusterLabel(orig_cluster = NULL, second_cluster = NULL, cores = 1)
}
\arguments{
\item{orig_cluster}{clustering label of first time}

\item{second_cluster}{clustering label of second time}

\item{cores}{parallel cores}
}
\description{
Map the K-means results to all cells
}
