% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingLouvain.R
\name{IntIndexCounting}
\alias{IntIndexCounting}
\title{Count internal index results}
\usage{
IntIndexCounting(sampling_list = NULL)
}
\arguments{
\item{sampling_list}{result from SamplingLouvain function}
}
\description{
It is a built-in function of SamplingLouvain for voting statistics of internal indicator results
}
