% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running.R
\name{kkl}
\alias{kkl}
\title{KKL clustering}
\usage{
kkl(
  dataMatrix = NULL,
  outlier_q = 0.2,
  down_n = 300,
  knn_range = 5:70,
  iter = 50,
  compute_index = c("Davies_Bouldin", "Calinski_Harabasz"),
  assess_index = "Davies_Bouldin",
  cores = 1,
  seed = 723
)
}
\arguments{
\item{seed}{}
}
\description{
This function is a wrapper that executes all steps of
KKL clustering analysis in one go.
}
