% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_clustering.R
\name{new_clustering}
\alias{new_clustering}
\title{Louvain/leiden clustering with optimal KNN structure}
\usage{
new_clustering(
  sampling_result = NULL,
  outlier_kmeans = NULL,
  assess_index = "Calinski_Harabasz",
  cluster_method = "louvain",
  res_range = seq(0.2, 3, 0.2),
  is_weight = TRUE,
  python_path = "/usr/bin/python3",
  seed = 723
)
}
\arguments{
\item{sampling_result}{sampling result from pre_partitioning function}

\item{outlier_kmeans}{outlier detection result from OutlierKmeans function}

\item{assess_index}{evaluation index used to select the optimal KNN graph structure}

\item{cluster_method}{louvain or leiden}

\item{res_range}{resolution range in clustering}

\item{python_path}{python path to be used}

\item{seed}{random seed}
}
\description{
This function is used to run Louvain clustering with optimal KNN graph structure
}
