% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResult.R
\name{NewLouvain}
\alias{NewLouvain}
\title{Louvain clustering with optimal KNN structure}
\usage{
NewLouvain(
  sampling_result = NULL,
  outlier_kmeans = NULL,
  assess_index = "Davies_Bouldin",
  cores = 1
)
}
\arguments{
\item{sampling_result}{sampling result from SamplingLouvain function}

\item{outlier_kmeans}{outlier detection result from OutlierKmeans function}

\item{assess_index}{evaluation index used to select the optimal KNN graph structure}

\item{cores}{the number of threads}
}
\description{
This function is used to run Louvain clustering with optimal KNN graph structure
}
