% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierKmeans.R
\name{OutlierDet}
\alias{OutlierDet}
\title{Outlier Detection}
\usage{
OutlierDet(m, q = 0.1, min_n = 3, cores = 1)
}
\arguments{
\item{m}{matrix, row is cells}

\item{q}{the proportion of outliers you want to filter}

\item{min_n}{the minimum number of cells required}

\item{cores}{parallel cores}
}
\value{
Character vector of outlier label
}
\description{
This function is used to detect outliers of any matrix.
}
