% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_partitioning.R
\name{OutlierDetED}
\alias{OutlierDetED}
\title{Outlier Detection}
\usage{
OutlierDetED(m, q = 0.2, min_n = 5)
}
\arguments{
\item{m}{matrix, row is cells}

\item{q}{the proportion of outliers you want to filter}

\item{min_n}{the minimum number of data points for outlier detection.}

\item{cores}{the number of threads}
}
\value{
Character vector of outlier label
}
\description{
This function is used to detect outliers of any matrix.
}
