% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_partitioning.R
\name{OutlierDetMD}
\alias{OutlierDetMD}
\title{Outlier Detection}
\usage{
OutlierDetMD(m, min_n = 5, alpha = 1, pvalue = 0.05)
}
\arguments{
\item{m}{matrix, row is cells}

\item{min_n}{the minimum number of cells required}

\item{alpha}{parameter of covMcd}

\item{pvalue}{significance level.}
}
\value{
Character vector of outlier label
}
\description{
This function is used to detect outliers of any matrix,
which comes from mvoutlier,
please check [github] (https://github.com/cran/mvoutlier) for more function details.
}
