% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierKmeans.R
\name{OutlierKmeans}
\alias{OutlierKmeans}
\title{K-means clustering based on outlier detection}
\usage{
OutlierKmeans(
  dataMatrix = NULL,
  down_n = 300,
  outlier_q = 0.2,
  cores = 1,
  seed = 723
)
}
\arguments{
\item{dataMatrix}{row is cell, col is feature}

\item{down_n}{downsampling number}

\item{outlier_q}{the proportion of outliers that need to be removed}

\item{cores}{the number of threads}

\item{seed}{random seed}
}
\value{
A list that include K-means result, division result, and center matrix was removed outliers
}
\description{
Get the result of K-means area division after outlier detection.
}
