% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{pcaMarchenkoPastur}
\alias{pcaMarchenkoPastur}
\title{Marchenko-Pastur Significant PCs (URD)}
\usage{
pcaMarchenkoPastur(M, N, pca.sdev, factor = 2, do.print = T)
}
\arguments{
\item{M}{(Numeric) Number of rows in input data, pc.gene}

\item{N}{(Numeric) Number of columns in input data, cell number}

\item{pca.sdev}{(Numeric vector) Standard deviations for each principal component}

\item{factor}{(Numeric) Factor to multiply eigenvalue null upper bound before determining significance.}

\item{do.print}{(Logical) Whether to report the results}
}
\value{
Logical vector of whether each PC is significant.
}
\description{
This function comes from URD, please check [github] (https://github.com/farrellja/URD) for more function details
}
