% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_partitioning.R
\name{pre_partitioning}
\alias{pre_partitioning}
\title{K-means clustering based on outlier detection}
\usage{
pre_partitioning(
  dat = NULL,
  partition_count = NULL,
  batch_size = min(500, nrow(dat)),
  num_init = 10,
  max_iters = 1000,
  outlier_det = TRUE,
  outlier_methods = "md",
  outlier_q = 0.2,
  min_n = 200,
  seed = 723
)
}
\arguments{
\item{dat}{expression matrix, row is cell, col is feature}

\item{partition_count}{the number of partitions}

\item{batch_size}{the number of threads}

\item{num_init}{the proportion of outliers that need to be removed}

\item{max_iters}{the maximum iterations for kmeans}

\item{outlier_det}{whether running outlier detection}

\item{outlier_methods}{outlier detection methods, md or ed.
"md" utilizes the Mahalanobis distance for detection,
while "ed" treats points that are far from the center proportionally as outliers.}

\item{outlier_q}{the proportion of outliers you want to filter.
When outlier_methods is ed, this parameter takes effect.}

\item{min_n}{the minimum number of data points for outlier detection.}

\item{seed}{random seed}
}
\value{
A list that include K-means result, division result, and center matrix was removed outliers
}
\description{
Get the result of K-means area division after outlier detection.
}
