% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingLouvain.R
\name{SamplingLouvain}
\alias{SamplingLouvain}
\title{Random sampling to do KNN-Louvain clustering}
\usage{
SamplingLouvain(
  dataMatrix = dataMatrix,
  outlier_kmeans = NULL,
  knn_range = c(3:70),
  iter = 30,
  compute_index = "Calinski_Harabasz",
  cores = 1,
  seed = NULL
)
}
\arguments{
\item{dataMatrix}{row is cell, col is feature}

\item{outlier_kmeans}{the result from OutlierKmeans function}

\item{knn_range}{the range of the number of neighbors in the KNN graph structure}

\item{iter}{the number of iterations}

\item{cores}{the number of threads}

\item{seed}{random seed}

\item{int_index}{the value of clustering quality index that need to be calculated}
}
\description{
This function is used to randomly sample each cluster multiple times
to select the most powerful KNN graph structure.
}
\details{
The currently available indices are:
\itemize{
   \item{Ball_Hall}\item{Banfeld_Raftery}\item{C_index}\item{Calinski_Harabasz}
   \item{Davies_Bouldin}\item{Det_Ratio}\item{Dunn}\item{Gamma}\item{G_plus}
   \item{Ksq_DetW}\item{Log_Det_Ratio}\item{Log_SS_Ratio}\item{McClain_Rao}
   \item{PBM}\item{Point_Biserial}\item{Ray_Turi}\item{Ratkowsky_Lance}
   \item{Scott_Symons}\item{SD_Scat}\item{SD_Dis}\item{S_Dbw}\item{Silhouette}
   \item{Tau}\item{Trace_W}\item{Trace_WiB}\item{Wemmert_Gancarski}
   \item{Xie_Beni}\item{GDI11}\item{GDI12}\item{GDI13}\item{GDI21}\item{GDI22}
   \item{GDI23}\item{GDI31}\item{GDI32}\item{GDI33}\item{GDI41}\item{GDI42}
   \item{GDI43}\item{GDI51}\item{GDI52}\item{GDI53}
}
}
