% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{stratified}
\alias{stratified}
\title{stratified sampling}
\usage{
stratified(df, group, size, select = NULL, replace = FALSE, bothSets = FALSE)
}
\arguments{
\item{df}{The input data.frame}

\item{group}{A character vector of the column or columns that make up the "strata".}

\item{size}{The desired sample size.
If size is a single integer of 1 or more, that number of samples is taken from each stratum.
If size is a vector of integers, the specified number of samples is taken for each stratum. It is recommended that you use a named vector. For example, if you have two strata, "A" and "B", and you wanted 5 samples from "A" and 10 from "B", you would enter size = c(A = 5, B = 10).}

\item{select}{This allows you to subset the groups in the sampling process. This is a list. For instance, if your group variable was "Group", and it contained three strata, "A", "B", and "C", but you only wanted to sample from "A" and "C", you can use select = list(Group = c("A", "C")).}

\item{replace}{For sampling with replacement.}
}
\description{
stratified sampling function from github：
https://gist.github.com/mrdwab/6424112
}
