% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingLouvain.R
\name{FindBestCluster}
\alias{FindBestCluster}
\title{Find best cluster result from internal indicators}
\usage{
FindBestCluster(data = NULL, meta_data = NULL, methods = c("all"), cores = 1)
}
\arguments{
\item{data}{row is cell, col is feature}

\item{meta_data}{columns represent different clustering results}

\item{methods}{index}

\item{cores}{parallel cores}
}
\description{
Find best cluster result from internal indicators
}
