% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running.R
\name{kkl}
\alias{kkl}
\title{KKL clustering}
\usage{
kkl(
  dataMatrix = NULL,
  outlier_q = 0.2,
  down_n = 300,
  knn_range = 5:70,
  iter = 50,
  compute_index = c("Davies_Bouldin", "Calinski_Harabasz"),
  assess_index = "Davies_Bouldin",
  cores = 1,
  seed = 723
)
}
\arguments{
\item{dataMatrix}{row is cell, col is feature}

\item{outlier_q}{the proportion of outliers that need to be removed}

\item{down_n}{downsampling number}

\item{knn_range}{the range of the number of neighbors in the KNN graph structure}

\item{iter}{the number of iterations}

\item{compute_index}{the value of clustering quality index that need to be calculated}

\item{assess_index}{evaluation index used to select the optimal KNN graph structure}

\item{cores}{the number of threads}

\item{seed}{random seed}
}
\description{
This function is a wrapper that executes all steps of
KKL clustering analysis in one go.
}
