% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetResult.R
\name{NewLouvain}
\alias{NewLouvain}
\title{Louvain clustering with optimal KNN structure}
\usage{
NewLouvain(
  sampling_result = NULL,
  outlier_kmeans = NULL,
  index = "Davies_Bouldin",
  cores = 1
)
}
\arguments{
\item{outlier_kmeans}{}
}
\description{
This function is used to run Louvain clustering with optimal KNN graph structure
}
